made the fetch report into a action/reducer
