+++ superset-frontend/src/dashboard/components/Header/index.jsx
@@ -20,13 +20,7 @@
 import moment from 'moment';
 import React from 'react';
 import PropTypes from 'prop-types';
-import rison from 'rison';
-import {
-  styled,
-  CategoricalColorNamespace,
-  t,
-  makeApi,
-} from '@superset-ui/core';
+import { styled, CategoricalColorNamespace, t } from '@superset-ui/core';
 import ButtonGroup from 'src/components/ButtonGroup';
 
 import {
@@ -406,34 +400,6 @@ class Header extends React.PureComponent {
     );
     return permissions[0].length > 0;
   }
-
-  fetchDashboardReport = async () => {
-    const { user, dashboardInfo } = this.props;
-    const queryParams = rison.encode({
-      filters: [
-        {
-          col: 'dashboard_id',
-          opr: 'eq',
-          value: dashboardInfo.id,
-        },
-        {
-          col: 'creation_method',
-          opr: 'eq',
-          value: 'dashboards',
-        },
-        {
-          col: 'created_by',
-          opr: 'rel_o_m',
-          value: user.userId,
-        },
-      ],
-    });
-    const response = await makeApi({
-      method: 'GET',
-      endpoint: '/api/v1/report',
-    })(`q=${queryParams}`);
-    this.setState({ attachedReports: response });
-  };
 
   render() {
     const {
